requirements.txt
# ATS-Optimizer Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching
redis==5.0.1

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# Optimization
cvxpy==1.4.2  # For linear/convex optimization

# Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Environment Variables
python-dotenv==1.0.0

# MQTT (for demand response)
paho-mqtt==1.6.1

# Logging
structlog==24.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0  # For testing async endpoints